requirements.txt
anthropic>=0.25.0
newsapi-python>=0.2.7
tweepy>=4.14.0
schedule>=1.2.1
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0
instagrapi>=2.0.0
psycopg[binary]>=3.1.0

# telephony-appointments (Riley) — FastAPI webhook server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
